requirements.txt
biopython
python-dotenv